refact: 공통 컴포넌트 분리 (필터, 검색창)
